move all routes to routes DIR
