+++ routes/index.js
@@ -1,6 +1,9 @@
 var data = require('./data.json');
 data.pkgVer = require('../package.json').version;
 
-exports.index = function(req, res){
-  res.render('index', data);
+module.exports = {
+  index: function(req, res){
+    res.render('index', data);
+  },
+  signup: require('./signup')
 };
